# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8081",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/tarea/obtenerOtis/{nombre}": {
      "get": {
        "tags": [
          "Tarea"
        ],
        "description": "Endpoint para obtener las otis del sector.",
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "nombre del sector"
          }
        ],
        "responses": {
          "200": {
            "description": "se obtiene otis",
            "schema": {
              "$ref": "#/definitions/arrayOti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/tarea/obtenerTareas/{id}": {
      "get": {
        "tags": [
          "Tarea"
        ],
        "description": "Endpoint para obtener las tareas asigandas de un operario.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID de oti"
          }
        ],
        "responses": {
          "200": {
            "description": "se obtiene tareas de un oti",
            "schema": {
              "$ref": "#/definitions/tareas"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/tarea/asignarTarea": {
      "post": {
        "tags": [
          "Tarea"
        ],
        "description": "Endpoint para asignar tarea a un operario.",
        "parameters": [
          {
            "name": "asignarTarea",
            "in": "body",
            "description": "datos para asignar una tarea",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/asignarTarea"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se obtiene tareas de un oti",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
...
    "/supervisor/": {
      "get": {
        "tags": [
          "Supervisor"
        ],
        "description": "Endpoint para obtener los supervisores.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "array de supervisores",
            "schema": {
              "$ref": "#/definitions/arraySupervisores"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
    "/supervisor/{id}": {
      "get": {
        "tags": [
          "Supervisor"
        ],
        "description": "Endpoint para obtener un supervisor.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Supervisor ID"
          }
        ],
        "responses": {
          "200": {
            "description": "SUpervisor encontrado",
            "schema": {
              "$ref": "#/definitions/supervisor"
            }
          },
          "404": {
            "description": "Error "
          }
        }
      }
    },
...
  },
  "definitions": {
    "asignarTarea": {
      "type": "object",
      "properties": {
        "fechaInicio": {
          "type": "string",
          "example": "2021/09/20"
        },
        "fechaFin": {
          "type": "string",
          "example": "2021/10/5"
        },
        "observacion": {
          "type": "string",
          "example": "obs_5"
        },
        "idOti": {
          "type": "string",
          "example": "616f2980c5e1d1846975c8cd"
        },
        "idOperario": {
          "type": "string",
          "example": "613e6361b2153ee73d786bc9"
        },
        "tareas": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "descripcion": {
                "type": "string",
                "example": "hola hola 3"
              },
              "nombre": {
                "type": "string",
                "example": "tarea nueva2"
              },
              "fechaInicio": {
                "type": "string",
                "example": "2009-11-11T03:00:00.000Z"
              },
              "fechaFin": {
                "type": "string",
                "example": "2009-11-15T03:00:00.000Z"
              },
              "tipoDeTarea": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "tarea5"
                  },
                  "descripcion": {
                    "type": "string",
                    "example": "descripcion_5"
                  },
                  "_id": {
                    "type": "string",
                    "example": "616f2981c5e1d1846975c8f6"
                  }
                }
              },
              "sector": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "deposito"
                  },
                  "operario": {
                    "type": "array",
                    "example": [],
                    "items": {}
                  },
                  "_id": {
                    "type": "string",
                    "example": "616f2981c5e1d1846975c8f7"
                  }
                }
              },
              "idOperario": {
                "type": "string",
                "example": "613e626ab2153ee73d77fb1d"
              },
              "estado": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "fechaInicio": {
                      "type": "string",
                      "example": "2009-11-11T03:00:00.000Z"
                    },
                    "fechaFin": {
                      "type": "string",
                      "example": "2009-11-15T03:00:00.000Z"
                    },
                    "observacion": {
                      "type": "string",
                      "example": "creado correctamente"
                    },
                    "tipoEstado": {
                      "type": "object",
                      "properties": {
                        "nombre": {
                          "type": "string",
                          "example": "iniciada"
                        },
                        "descripcion": {
                          "type": "string",
                          "example": "se inicializo correctamente"
                        },
                        "_id": {
                          "type": "string",
                          "example": "616f2981c5e1d1846975c8f9"
                        }
                      }
                    },
                    "_id": {
                      "type": "string",
                      "example": "616f2981c5e1d1846975c8f8"
                    }
                  }
                }
              },
              "_id": {
                "type": "string",
                "example": "616f2981c5e1d1846975c8f5"
              }
            }
          }
        }
      }
    },
...
    "oti": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "613e5777b2153ee73d73142f"
        },
        "fechaFin": {
          "type": "string",
          "example": ""
        },
        "rosca": {
          "type": "object",
          "properties": {
            "descripcionTecnica": {
              "type": "string",
              "example": "es una rosca dificil de hacer"
            },
            "medida": {
              "type": "string",
              "example": "20x50x10"
            },
            "tipoDeRosca": {
              "type": "object",
              "properties": {
                "descripcion": {
                  "type": "string",
                  "example": "la mejor rosca"
                },
                "categoria": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "premium"
                    },
                    "descripcion": {
                      "type": "string",
                      "example": "exportacion marina y utilizacion para caños de gas"
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e53c8db8fc5cc55e1fb3d"
                    }
                  }
                },
                "_id": {
                  "type": "string",
                  "example": "615e53c8db8fc5cc55e1fb3c"
                }
              }
            },
            "_id": {
              "type": "string",
              "example": "615e53c8db8fc5cc55e1fb3b"
            }
          }
        },
        "sector": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "operario": {
                "type": "array",
                "example": [],
                "items": {}
              },
              "_id": {
                "type": "string",
                "example": "61a382527d3c127a35ddf813"
              },
              "nombre": {
                "type": "string",
                "example": "deposito"
              }
            }
          }
        },
        "tareas": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "descripcion": {
                "type": "string",
                "example": "descripcion_2"
              },
              "nombre": {
                "type": "string",
                "example": "tarea_2"
              },
              "fechaInicio": {
                "type": "string",
                "example": "2020-03-02T03:00:00.000Z"
              },
              "fechaFin": {
                "type": "string",
                "example": "2020-01-04T03:00:00.000Z"
              },
              "numeroDeOrden": {
                "type": "number",
                "example": 2
              },
              "tipoDeTarea": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "tarea3"
                  },
                  "descripcion": {
                    "type": "string",
                    "example": "descripcion_3"
                  },
                  "_id": {
                    "type": "string",
                    "example": "615e53c8db8fc5cc55e1fb42"
                  }
                }
              },
              "sector": {
                "type": "object",
                "properties": {
                  "operario": {
                    "type": "array",
                    "example": [],
                    "items": {}
                  },
                  "nombre": {
                    "type": "string",
                    "example": "deposito"
                  },
                  "_id": {
                    "type": "string",
                    "example": "615e53c8db8fc5cc55e1fb43"
                  }
                }
              },
              "estado": {
                "type": "array",
                "example": [],
                "items": {}
              },
              "_id": {
                "type": "string",
                "example": "615e53c8db8fc5cc55e1fb41"
              }
            }
          }
        },
        "estados": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "fechaInicio": {
                "type": "string",
                "example": "2021-09-20T03:00:00.000Z"
              },
              "fechaFin": {
                "type": "string",
                "example": "2021-10-05T03:00:00.000Z"
              },
              "observacion": {
                "type": "string",
                "example": "obs_5"
              },
              "tipoEstado": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "Iniciada"
                  },
                  "descripcion": {
                    "type": "string",
                    "example": ""
                  },
                  "_id": {
                    "type": "string",
                    "example": "615e5b1e2968697488389e79"
                  }
                }
              },
              "_id": {
                "type": "string",
                "example": "615e5b1e2968697488389e78"
              }
            }
          }
        }
      }
    },
    "arrayOti": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "613e5777b2153ee73d73142f"
          },
          "fechaFin": {
            "type": "string",
            "example": ""
          },
          "rosca": {
            "type": "object",
            "properties": {
              "descripcionTecnica": {
                "type": "string",
                "example": "es una rosca dificil de hacer"
              },
              "medida": {
                "type": "string",
                "example": "20x50x10"
              },
              "tipoDeRosca": {
                "type": "object",
                "properties": {
                  "descripcion": {
                    "type": "string",
                    "example": "la mejor rosca"
                  },
                  "categoria": {
                    "type": "object",
                    "properties": {
                      "nombre": {
                        "type": "string",
                        "example": "premium"
                      },
                      "descripcion": {
                        "type": "string",
                        "example": "exportacion marina y utilizacion para caños de gas"
                      },
                      "_id": {
                        "type": "string",
                        "example": "615e53c8db8fc5cc55e1fb3d"
                      }
                    }
                  },
                  "_id": {
                    "type": "string",
                    "example": "615e53c8db8fc5cc55e1fb3c"
                  }
                }
              },
              "_id": {
                "type": "string",
                "example": "615e53c8db8fc5cc55e1fb3b"
              }
            }
          },
          "sector": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "operario": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "_id": {
                  "type": "string",
                  "example": "61a382527d3c127a35ddf813"
                },
                "nombre": {
                  "type": "string",
                  "example": "deposito"
                }
              }
            }
          },
          "tareas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "descripcion": {
                  "type": "string",
                  "example": "descripcion_2"
                },
                "nombre": {
                  "type": "string",
                  "example": "tarea_2"
                },
                "fechaInicio": {
                  "type": "string",
                  "example": "2020-03-02T03:00:00.000Z"
                },
                "fechaFin": {
                  "type": "string",
                  "example": "2020-01-04T03:00:00.000Z"
                },
                "numeroDeOrden": {
                  "type": "number",
                  "example": 2
                },
                "tipoDeTarea": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "tarea3"
                    },
                    "descripcion": {
                      "type": "string",
                      "example": "descripcion_3"
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e53c8db8fc5cc55e1fb42"
                    }
                  }
                },
                "sector": {
                  "type": "object",
                  "properties": {
                    "operario": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "nombre": {
                      "type": "string",
                      "example": "deposito"
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e53c8db8fc5cc55e1fb43"
                    }
                  }
                },
                "estado": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "_id": {
                  "type": "string",
                  "example": "615e53c8db8fc5cc55e1fb41"
                }
              }
            }
          },
          "estados": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fechaInicio": {
                  "type": "string",
                  "example": "2021-09-20T03:00:00.000Z"
                },
                "fechaFin": {
                  "type": "string",
                  "example": "2021-10-05T03:00:00.000Z"
                },
                "observacion": {
                  "type": "string",
                  "example": "obs_5"
                },
                "tipoEstado": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "Iniciada"
                    },
                    "descripcion": {
                      "type": "string",
                      "example": ""
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e5b1e2968697488389e79"
                    }
                  }
                },
                "_id": {
                  "type": "string",
                  "example": "615e5b1e2968697488389e78"
                }
              }
            }
          }
        }
      }
    },
...
    "supervisor": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "613bc8ccc7d979667ca2a137"
        },
        "numeroEmpleado": {
          "type": "string",
          "example": "2"
        },
        "DNI": {
          "type": "string",
          "example": "66316022"
        },
        "nombre": {
          "type": "string",
          "example": "martin"
        },
        "apellido": {
          "type": "string",
          "example": "garcia"
        },
        "telefono": {
          "type": "string",
          "example": "850-421-9067"
        },
        "nombreUsuario": {
          "type": "string",
          "example": "supervisor1234"
        },
        "contraseña": {
          "type": "string",
          "example": "12345"
        },
        "email": {
          "type": "string",
          "example": "martingarcia@gmail.com"
        }
      }
    },
    "arraySupervisores": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "613bc8ccc7d979667ca2a137"
          },
          "numeroEmpleado": {
            "type": "string",
            "example": "2"
          },
          "DNI": {
            "type": "string",
            "example": "66316022"
          },
          "nombre": {
            "type": "string",
            "example": "martin"
          },
          "apellido": {
            "type": "string",
            "example": "garcia"
          },
          "telefono": {
            "type": "string",
            "example": "850-421-9067"
          },
          "nombreUsuario": {
            "type": "string",
            "example": "supervisor1234"
          },
          "contraseña": {
            "type": "string",
            "example": "12345"
          },
          "email": {
            "type": "string",
            "example": "martingarcia@gmail.com"
          }
        }
      }
    },
...
    "tareas": {
      "type": "object",
      "properties": {
        "tareasOperario": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "descripcion": {
                "type": "string",
                "example": "asignar a operario con experiencia"
              },
              "nombre": {
                "type": "string",
                "example": "cortar pieza"
              },
              "fechaInicio": {
                "type": "string",
                "example": "YYYY/MM/DD"
              },
              "fechaFin": {
                "type": "string",
                "example": "YYYY/MM/DD"
              },
              "numeroDeOrden": {
                "type": "string",
                "example": "4589"
              },
              "tipoDeTarea": {
                "type": "string",
                "example": "tareas de corte"
              },
              "sector": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "corte"
                  },
                  "activo": {
                    "type": "string",
                    "example": "true"
                  },
                  "operario": {
                    "type": "array",
                    "example": [
                      "5ea7029320e515102e44c71d",
                      "5ea7029320e515102ed5c22d",
                      "5ea7029320e515222102ed5c71d",
                      "5ea7029320e515123pc71d"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "idOperario": {
                "type": "string",
                "example": "5ea7029320e515102e44c71d"
              },
              "estado": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "fechaInicio": {
                      "type": "string",
                      "example": "YYYY/MM/DD"
                    },
                    "fechaFin": {
                      "type": "string",
                      "example": "YYYY/MM/DD"
                    },
                    "observacion": {
                      "type": "string",
                      "example": "si inicia con todos las herramientas calibradas"
                    },
                    "tipoEstado": {
                      "type": "object",
                      "properties": {
                        "nombre": {
                          "type": "string",
                          "example": "en progreso"
                        },
                        "descripcion": {
                          "type": "string",
                          "example": "en proceso de produccion"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "idOti": {
          "type": "string",
          "example": "616f2980c5e1d1846975c8cd"
        }
      }
    },
...
  }
}